Move the alias resoloving logic to dedicated classes
