+++ packages/ballerina-language-server/flow-model-generator/modules/flow-model-generator-ls-extension/src/main/java/io/ballerina/flowmodelgenerator/extension/ExpressionEditorService.java
@@ -46,7 +46,7 @@ import io.ballerina.flowmodelgenerator.extension.response.FunctionCallTemplateRe
 import io.ballerina.flowmodelgenerator.extension.response.ImportModuleResponse;
 import io.ballerina.flowmodelgenerator.extension.response.VisibleVariableTypesResponse;
 import io.ballerina.modelgenerator.commons.CommonUtils;
-import io.ballerina.modelgenerator.commons.ModuleAliasResolver;
+import io.ballerina.modelgenerator.commons.ImportPrefixReader;
 import io.ballerina.modelgenerator.commons.ModuleInfo;
 import io.ballerina.modelgenerator.commons.PackageUtil;
 import io.ballerina.projects.Document;
@@ -342,7 +342,7 @@ public class ExpressionEditorService implements ExtendedLanguageServerService {
                     .orElse(null);
             // A null root falls back to the module's natural prefix, which is what this method promises without a
             // file to read. An import is emitted for the result, so allocating a free prefix is the right answer.
-            return ModuleAliasResolver.resolve(rootNode, codedata.org(), codedata.module(), null);
+            return ImportPrefixReader.resolve(rootNode, codedata.org(), codedata.module(), null);
         } catch (RuntimeException e) {
             // Without a file to read, the module's natural prefix is the only available answer.
             return codedata.getModulePrefix();
